+++ CHANGELOG.md
@@ -30,6 +30,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Use bindata as the default method of accessing templates, this prevents many
   bug reports we've had in the past. During development or to otherwise opt out
   the --basedir flag will not load from bindata (driver's are the exception).
+- Export the queries.BuildQuery method for public use. This allows building
+  queries without executing.
 
 ### Changed